+++ packages/plugin-compiler/src/test/kotlin/io/realm/kotlin/compiler/GenerationExtensionTest.kt
@@ -34,6 +34,7 @@ import io.realm.kotlin.internal.schema.PropertyMetadata
 import io.realm.kotlin.internal.schema.SchemaMetadata
 import io.realm.kotlin.types.BaseRealmObject
 import io.realm.kotlin.types.RealmObject
+import io.realm.kotlin.types.TypedRealmObject
 import org.junit.Test
 import java.io.File
 import kotlin.reflect.KClass
@@ -373,6 +374,8 @@ class GenerationExtensionTest {
                         get() = TODO("Not yet implemented")
                     override val properties: List<PropertyMetadata>
                         get() = TODO("Not yet implemented")
+                    override val clazz: KClass<out TypedRealmObject>?
+                        get() = TODO("Not yet implemented")
                     override val className: String
                         get() = TODO("Not yet implemented")
                     override val primaryKeyProperty: PropertyMetadata?
@@ -388,6 +391,10 @@ class GenerationExtensionTest {
                     override fun get(propertyName: String): PropertyMetadata? {
                         TODO("Not yet implemented")
                     }
+                }
+
+                override fun get(classKey: ClassKey): ClassMetadata? {
+                    TODO("Not yet implemented")
                 }
             }
     }