+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This release will bump the Realm file format from version 22 to 23. Opening a fi
 * [Sync] `App.close()` have been added so it is possible to close underlying ressources used by the app instance.
 * [Sync] Add support for progress listeners with `SyncSession.progress`. (Issue [#428](https://github.com/realm/realm-kotlin/issues/428))
 * Add better error messages when inheriting `RealmObject` with unsupported class types. (Issue [#1086](https://github.com/realm/realm-kotlin/issues/1086))
+* Added support for reverse relationships on Embedded objects through the `EmbeddedRealmObject.parent()` extension function. (Issue [#1141](https://github.com/realm/realm-kotlin/pull/1141))
+* Added support for reverse relationships through the `backlinks` delegate on `EmbeddedObjects`. See the function documentation for more details. (Issue [#1134](https://github.com/realm/realm-kotlin/issues/1134))
 * Added support for `@PersistedName` annotations for mapping a Kotlin field name to the underlying field name persisted in the Realm. (Issue [#590](https://github.com/realm/realm-kotlin/issues/590))
 
 ### Fixed
@@ -54,6 +56,7 @@ This release will bump the Realm file format from version 22 to 23. Opening a fi
 
 ### Fixed
 * Added proguard consumer files to Android debug artifacts. (Issue [#1150](https://github.com/realm/realm-kotlin/issues/1150))
+* Fixed bug when creating `RealmInstant` instaces with `RealmInstant.now()` in Kotlin Native. (Issue [#1182](https://github.com/realm/realm-kotlin/issues/1182))
 
 ### Compatibility
 * This release is compatible with the following Kotlin releases:
@@ -67,7 +70,7 @@ This release will bump the Realm file format from version 22 to 23. Opening a fi
 * Minimum Android SDK: 16.
 
 ### Internal
-* None.
+* Updated to Gradle 7.6.
 
 
 ## 1.5.1 (2022-12-12)